requirements.txt
PyQt5>=5.15.0
torch>=2.0.0
transformers>=4.30.0
python-docx>=0.8.11
PyPDF2>=3.0.0
striprtf>=0.0.26
reportlab>=4.0.0
biopython>=1.81
